% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{babynamesIL_totals}
\alias{babynamesIL_totals}
\title{Israeli baby names total numbers.}
\format{
A data frame with 4 variables: \code{sector}, \code{sex}, \code{name}, \code{n}.
}
\usage{
babynamesIL_totals
}
\description{
Total number of babies per name, sector and gender at the years 1948-2023.
}
\keyword{datasets}
